Refactor Loader load method
Moves some of the logic from the Loader class to the Mapper class.
This is prep work for generalizing mapper support to make
it super easy to add new mappers.
