+++ GhidraNes/src/main/java/ghidranes/GhidraNesLoader.java
@@ -17,7 +17,9 @@ package ghidranes;
 
 import java.io.IOException;
 import java.io.InputStream;
-import java.util.*;
+import java.util.ArrayList;
+import java.util.Collection;
+import java.util.List;
 
 import ghidra.app.util.Option;
 import ghidra.app.util.bin.ByteProvider;
@@ -26,19 +28,11 @@ import ghidra.app.util.opinion.AbstractLibrarySupportLoader;
 import ghidra.app.util.opinion.LoadSpec;
 import ghidra.framework.model.DomainObject;
 import ghidra.framework.store.LockException;
-import ghidra.program.model.address.AddressOutOfBoundsException;
 import ghidra.program.model.address.AddressOverflowException;
-import ghidra.program.model.address.AddressSpace;
 import ghidra.program.model.lang.Language;
 import ghidra.program.model.lang.LanguageCompilerSpecPair;
 import ghidra.program.model.listing.Program;
-import ghidra.program.model.address.Address;
-import ghidra.program.model.mem.Memory;
-import ghidra.program.model.mem.MemoryAccessException;
 import ghidra.program.model.mem.MemoryConflictException;
-import ghidra.program.model.symbol.SourceType;
-import ghidra.program.model.symbol.Symbol;
-import ghidra.program.model.symbol.SymbolTable;
 import ghidra.util.exception.CancelledException;
 import ghidra.util.exception.DuplicateNameException;
 import ghidra.util.exception.InvalidInputException;
@@ -47,7 +41,6 @@ import ghidranes.errors.NesRomException;
 import ghidranes.errors.UnimplementedNesMapperException;
 import ghidranes.mappers.NesMapper;
 import ghidranes.util.MemoryBlockDescription;
-import ghidranes.util.NesMmio;
 
 /**
  * This loader parses an iNES ROM file and maps the PRG and CHR rom appropriately
@@ -97,86 +90,11 @@ public class GhidraNesLoader extends AbstractLibrarySupportLoader {
 		NesMapper mapper;
 		try {
 			mapper = NesMapper.getMapper(rom.header.mapper);
-			mapper.updateMemoryMapForRom(rom, program, monitor);
-		} catch (LockException | MemoryConflictException | AddressOverflowException | DuplicateNameException | UnimplementedNesMapperException e) {
+			mapper.apply(rom, program, monitor);
+		} catch (LockException | MemoryConflictException | AddressOverflowException
+				 | DuplicateNameException | InvalidInputException | UnimplementedNesMapperException e) {
 			throw new RuntimeException(e);
 		}
-
-		try {
-			// RES should have the highest precedence, followed by NMI, followed by IRQ. We set them
-			// as primary in reverse order because the last `.setPrimary()` call has precedence
-			addVectorEntryPoint(program, "IRQ", 0xFFFE, "irq");
-			addVectorEntryPoint(program, "NMI", 0xFFFA, "vblank");
-			addVectorEntryPoint(program, "RES", 0xFFFC, "reset");
-
-			AddressSpace addressSpace = program.getAddressFactory().getDefaultAddressSpace();
-			SymbolTable symbolTable = program.getSymbolTable();
-
-			ArrayList<NesMmio> registers = new ArrayList<NesMmio>(Arrays.asList(
-				new NesMmio(addressSpace, 0x2000, "PPUCTRL"),
-				new NesMmio(addressSpace, 0x2001, "PPUMASK"),
-				new NesMmio(addressSpace, 0x2002, "PPUSTATUS"),
-				new NesMmio(addressSpace, 0x2003, "OAMADDR"),
-				new NesMmio(addressSpace, 0x2004, "OAMDATA"),
-				new NesMmio(addressSpace, 0x2005, "PPUSCROLL"),
-				new NesMmio(addressSpace, 0x2006, "PPUADDR"),
-				new NesMmio(addressSpace, 0x2007, "PPUDATA"),
-				new NesMmio(addressSpace, 0x4000, "SQ1_VOL"),
-				new NesMmio(addressSpace, 0x4001, "SQ1_SWEEP"),
-				new NesMmio(addressSpace, 0x4002, "SQ1_LO"),
-				new NesMmio(addressSpace, 0x4003, "SQ1_HI"),
-				new NesMmio(addressSpace, 0x4004, "SQ2_VOL"),
-				new NesMmio(addressSpace, 0x4005, "SQ2_SWEEP"),
-				new NesMmio(addressSpace, 0x4006, "SQ2_LO"),
-				new NesMmio(addressSpace, 0x4007, "SQ2_HI"),
-				new NesMmio(addressSpace, 0x4008, "TRI_LINEAR"),
-				new NesMmio(addressSpace, 0x400a, "TRI_LO"),
-				new NesMmio(addressSpace, 0x400b, "TRI_HI"),
-				new NesMmio(addressSpace, 0x400c, "NOISE_VOL"),
-				new NesMmio(addressSpace, 0x400e, "NOISE_LO"),
-				new NesMmio(addressSpace, 0x400f, "NOISE_HI"),
-				new NesMmio(addressSpace, 0x4010, "DMC_FREQ"),
-				new NesMmio(addressSpace, 0x4011, "DMC_RAW"),
-				new NesMmio(addressSpace, 0x4012, "DMC_START"),
-				new NesMmio(addressSpace, 0x4013, "DMC_LEN"),
-				new NesMmio(addressSpace, 0x4014, "OAMDMA"),
-				new NesMmio(addressSpace, 0x4015, "SND_CHN"),
-				new NesMmio(addressSpace, 0x4016, "JOY1"),
-				new NesMmio(addressSpace, 0x4017, "JOY2")
-			));
-
-			mapper.addExtraRegisters(registers, addressSpace);
-
-			for (NesMmio register : registers) {
-				Symbol s = symbolTable.createLabel(register.address, register.name, SourceType.IMPORTED);
-				s.setPinned(true);
-			}
-		} catch (InvalidInputException | AddressOutOfBoundsException | MemoryAccessException e) {
-			throw new RuntimeException(e);
-		}
-	}
-
-	protected void addVectorEntryPoint(Program program, String vectorLabel, long vectorAddress, String targetLabel)
-		 throws InvalidInputException, MemoryAccessException {
-		AddressSpace addressSpace = program.getAddressFactory().getDefaultAddressSpace();
-		SymbolTable symbolTable = program.getSymbolTable();
-		Memory memory =  program.getMemory();
-
-		Address vAddress = addressSpace.getAddress(vectorAddress);
-		Symbol vSymbol = symbolTable.createLabel(vAddress, vectorLabel, SourceType.IMPORTED);
-		vSymbol.setPinned(true);
-		vSymbol.setPrimary();
-		symbolTable.addExternalEntryPoint(vAddress);
-
-		byte vecLo = memory.getByte(vAddress);
-		byte vecHi = memory.getByte(vAddress.add(1));
-		long vec = (Byte.toUnsignedLong(vecHi) << 8) | Byte.toUnsignedLong(vecLo);
-		Address vTargetAddress = addressSpace.getAddress(vec);
-
-		Symbol vTargetSymbol = symbolTable.createLabel(vTargetAddress, targetLabel, SourceType.IMPORTED);
-		symbolTable.addExternalEntryPoint(vTargetAddress);
-
-		vTargetSymbol.setPrimary();
 	}
 
 	@Override
@@ -197,7 +115,7 @@ public class GhidraNesLoader extends AbstractLibrarySupportLoader {
 	protected void createDefaultMemoryBlocks(Program program, Language language, MessageLog log) {
 		// NOTE: We skip the default memory blocks because Ghidra's default 6502 memory map
 		// differs from the NES's memory map
-//		super.createDefaultMemoryBlocks(program, language, log);
+		//		super.createDefaultMemoryBlocks(program, language, log);
 
 		try {
 			int ramPermissions =
